+++ Project3/Controllers/InfoStudentController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Project3.Entity.Request;
 using Project3.Migrations;
@@ -16,7 +17,7 @@ namespace Project3.Controllers
             _infoService = infoService;
         }
 
-        [HttpPost("update")]
+        [HttpPost("update"), Authorize]
         public async Task<IActionResult> createOrUpdate([FromBody] UpdateInformationStudent? menuReq)
         {
             return Ok(_infoService.UpdateInfo(menuReq));
@@ -28,13 +29,13 @@ namespace Project3.Controllers
             return Ok(_infoService.getOne(Id));
         }
 
-        [HttpDelete("deleteById/{id}")]
+        [HttpDelete("deleteById/{id}") , Authorize]
         public async Task<IActionResult> deleteById(long id)
         {
             return Ok(_infoService.deleteAcount(id));
         }
 
-        [HttpPost("search")]
+        [HttpPost("search") , Authorize]
         public async Task<IActionResult> search([FromBody] InfomationStudentReq filter)
         {
             return Ok(_infoService.getPagin(filter));
